+++ src/app/routes.ts
@@ -11,7 +11,9 @@ export const routes: Routes = [
   {
     path: 'tax-report/:taxReportId',
     loadComponent: () =>
-      import('./payment/components/payment-shell').then((mod) => mod.PaymentShellComponent),
+      import('./payment/components/payment-shell/payment-shell.component').then(
+        (mod) => mod.PaymentShellComponent
+      ),
   },
   {
     path: '',
